unpack a buffer and invoke a functor
Added an unpack_and_invoke method to unpack a transparently_serializable buffer into a tuple of references and call a an object supplied using std::apply. Objects are created in-place in the supplied buffer which must be suitably aligned for the types in the pack. returns true if F were invoked, and false otherwise.